Backend/requirements.txt
asyncio
aiohttp
google-cloud-videointelligence
SendGrid
nltk
numpy